website_files/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
pandas==0.25.0
numpy==1.14.3
scikit-learn==0.21.2
flask_restful
scipy
google-cloud-storage==1.13.2
